+++ crates/serialization/tests/integration/main.rs
@@ -15,6 +15,7 @@
 
 mod test_enums_capnp;
 mod test_identifiers_capnp;
+mod test_instruments_capnp;
 mod test_market_data_capnp;
 mod test_market_data_sbe;
 mod test_types_capnp;